requirements.txt
yt-dlp>=2025.1.26
pytube>=15.0.0
youtube-comment-downloader>=0.1.70
gdown>=5.2.0
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
jinja2>=3.1.4
fastapi>=0.115.0
uvicorn>=0.30.0
Flask>=3.0.3